Animal-image-classification/requirements.txt
flask
tensorflow
Pillow
numpy 